Backend/requirements.txt
click==8.1.3
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
requests==2.32.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
Pillow==11.2.1
numpy==2.1.3
opencv-python==4.11.0.86
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.1
python-dotenv==1.1.0
Flask-Cors==4.0.0
